backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.110.1
h11==0.14.0
idna==3.7
pydantic==2.6.4
pydantic-core==2.16.3
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.37.2
typing-extensions==4.11.0
uvicorn==0.29.0
